pyproject.toml
[project]
name = "rag-model"
version = "0.1.0"
description = "my model app, to get insights and chat your questions and tasks"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "chromadb>=0.5.0",
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "langchain-core>=0.2.9",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "ollama>=0.3.0",
    "pypdf>=4.2.0",
    "sentence-transformers>=2.7.0",
    "fastapi[standard]>=0.111.0",
    "uvicorn>=0.29.0",
    "pydantic-settings>=2.3.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "bcrypt>=4.0",
    "pyjwt[crypto]>=2.8",
    "alembic>=1.13",
    "email-validator>=2.0",
    "huggingface-hub>=0.30",
    "celery>=5.6.2",
    "redis>=7.2.0",
    "fastapi-limiter>=0.2.0",
    "uuid6>=2025.0.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
addopts = "-m 'not integration'"
markers = [
    "email_integration: mark test as requiring email integration (requires Gmail credentials)",
    "integration: mark test as an integration test requiring the full stack (ChromaDB, Ollama, Postgres)",
]
